detect notification in customer app
